+++ Database/DatabaseConfig.cs
@@ -1,0 +1,6 @@
+namespace LabManager.Database;
+
+class DatabaseConfig 
+{
+    public string ConnectionString { get => "Data Source=database.db"; }
+}
