+++ Database/DatabaseSetup.cs
@@ -4,19 +4,21 @@ namespace LabManager.Database;
 
 class DatabaseSetup
 {
-    public DatabaseSetup()
+    private DatabaseConfig databaseConfig;
+
+    public DatabaseSetup(DatabaseConfig databaseConfig)
     {
+        this.databaseConfig = databaseConfig;
         CreateTableComputer();
-        CreateTableLab();
     }
 
     public void CreateTableComputer()
     {
-        var connection = new SqliteConnection("Data Source = database.db");
-
+        var connection = new SqliteConnection(databaseConfig.ConnectionString);
         connection.Open();
 
         var command = connection.CreateCommand();
+        
         command.CommandText = @"
             CREATE TABLE IF NOT EXISTS Computers(
                 id int not null primary key,
