+++ Repositories/ComputerRepository.cs
@@ -1,3 +1,4 @@
+using LabManager.Database;
 using LabManager.Models;
 using Microsoft.Data.Sqlite;
 
@@ -5,6 +6,9 @@ namespace LabManager.Repositories;
 
 class ComputerRepository
 {
+    private DatabaseConfig databaseConfig;
+
+    public ComputerRepository(DatabaseConfig databaseConfig) => this.databaseConfig = databaseConfig;
     public List<Computer> GetAll()
     {
         var computers = new List<Computer>();
@@ -27,5 +31,21 @@ class ComputerRepository
         connection.Close();
 
         return computers;
+
+    }
+
+    public void Save(Computer computer)
+    {
+        var connection = new SqliteConnection("Data Source = database.db");
+        connection.Open();
+
+        var command = connection.CreateCommand();
+        command.CommandText = "INSERT INTO Computers VALUES($id, $ram, $processor);";
+        command.Parameters.AddWithValue("$id", computer.Id);
+        command.Parameters.AddWithValue("$ram", computer.Ram);
+        command.Parameters.AddWithValue("$processor", computer.Processor);
+
+        command.ExecuteNonQuery();
+        connection.Close();
     }
 }
